Fix fields of first article when running init
